requirements.txt
fastapi
uvicorn
pydantic
numpy
torch
torchvision
Pillow
opencv-python
tqdm
pyyaml
